+++ invitation-domain/src/main/java/com/nailseong/invitation/invitation/appilication/domain/InvitationRepository.java
@@ -1,0 +1,6 @@
+package com.nailseong.invitation.invitation.appilication.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface InvitationRepository extends JpaRepository<Invitation, Long> {
+}
