--- invitation-infrastructure/src/main/java/com/nailseong/invitation/invitation/InvitationRepository.java
@@ -1,6 +1,0 @@
-package com.nailseong.invitation.invitation;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface InvitationRepository extends JpaRepository<InvitationEntity, Long> {
-}
